requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
pypdf>=4.0.0
pillow>=10.1.0
pillow-heif>=0.16.0
reportlab>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
